+++ copier/errors.py
@@ -109,3 +109,7 @@ class UnknownCopierVersionWarning(UserWarning, CopierWarning):
 
 class OldTemplateWarning(UserWarning, CopierWarning):
     """Template was designed for an older Copier version."""
+
+
+class DirtyLocalWarning(UserWarning, CopierWarning):
+    """Changes and untracked files present in template."""