+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Summary:
 -   Raise a CopierAnswersInterrupt instead of a bare KeyboardInterrupt to provide
     callers with additional context - such as the partially completed AnswersMap.
 -   Support for `user_defaults`, which take precedence over template defaults.
+-   Copy dirty changes from a git-tracked template to the project by default, to make
+    testing easier.
 
 ### Changed